+++ src/main.rs
@@ -5,6 +5,5 @@ use structopt::StructOpt;
 
 fn main() -> Result<()> {
     let opt = Opt::from_args();
-    #[allow(clippy::cast_possible_truncation)]
-    process::exit(process_files(&opt)? as i32)
+    process::exit(process_files(&opt)?)
 }
